backend/requirements.txt
fastapi
uvicorn
openai
google-cloud-vision
aiohttp
aiofiles
asyncpg
python-multipart
pydantic
requests
Pillow
uvicorn main:app --host 0.0.0.0 --port 10000